+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Expose justification mechanism for TRLC files (up, down, and
   global).
 
+* Improve stats display in HTML report.
+
 ### 0.9.9
 
 * In the HTML report, items are now sorted in a meaningful way.